+++ src/components/icons/Icon.tsx
@@ -4,11 +4,15 @@ import type { AnyComponent } from "../../types";
 import { Checkmark } from "./Checkmark.js";
 import { Back } from "./Back.js";
 import { Forward } from "./Forward.js";
+import { Search } from "./Search.js";
+import { Battery } from "./Battery.js";
 
 enum ICON {
     checkmark = "checkmark",
     back = "back",
     forward = "forward",
+    search = "search",
+    battery = "battery",
 }
 
 interface Props {
@@ -20,6 +24,8 @@ const iconsMap = new Map<ICON, AnyComponent>([
     [ICON.checkmark, Checkmark],
     [ICON.back, Back],
     [ICON.forward, Forward],
+    [ICON.search, Search],
+    [ICON.battery, Battery],
 ]);
 
 const Icon: React.FC<Props> = ({ name: iconName, ...props }) => {