+++ src/index.ts
@@ -1,4 +1,6 @@
 export * from "./components/form/Checkbox.js";
+export * from "./components/form/Input.js";
+export * from "./components/form/Label.js";
 export * from "./components/form/Toggle.js";
 
 export * from "./components/icons/Icon.js";
@@ -9,6 +11,7 @@ export * from "./components/layout/header/HeaderIconAction.js";
 export * from "./components/layout/header/StickyHeader.js";
 export * from "./components/layout/list/List.js";
 export * from "./components/layout/list/Item.js";
+export * from "./components/layout/section/SearchContainer.js";
 export * from "./components/layout/section/Section.js";
 
 export * from "./components/ui/action/Action.js";